requirements.txt
bitstring
matplotlib
librosa>=0.7.2
numpy>=1.23.4
pandas>=1.1.5
Pillow>=9.0.1
praat-parselmouth
pydub
pyloudnorm>=0.1.0
pyrubberband
PyWavelets>=1.1.1
scikit-learn>=0.22
SoundFile>=0.10.3.post1
sox>=1.4.0
tabulate>=0.8.9
tqdm>=4.61.2
tensorboard>=2.10.1
pyYaml>=5.4.1
pydub
torch==2.1.1
torchaudio==2.1.1